+++ crates/ruma-client-api/CHANGELOG.md
@@ -42,6 +42,9 @@ Breaking changes:
   change.
 - The endpoints that were using the `NoAuthentication` authentication scheme now
   use `NoAccessToken`.
+- The variants of `UserIdentifier` were changed to tuple variants containing a
+  non-exhaustive struct.
+  - `UserIdentifier::UserIdOrLocalpart` was renamed to `UserIdentifier::Matrix`.
 
 Bug fixes: